pyproject.toml
[project]
name = "chatboot-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.54.3",
    "psycopg2>=2.9.10",
    "pypika>=0.48.9",
    "python-dotenv>=1.0.1",
    "reportlab>=4.2.5",
    "streamlit>=1.40.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
